refactor(promql): let the range function consume shard sources directly
The range function does not need its own plan: shard_sources in
series_source::stream turns a selector into one StreamSource per shard,
the fold and the emitter consume those, and the engine composes the two.
fused::stream keeps only the fold entry.
